backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
feedparser==6.0.11
PyYAML==6.0.2
APScheduler==3.10.4
pytz==2024.1
